% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near_prime.R
\name{near_prime}
\alias{near_prime}
\title{'Find the Nearest Pirme Number'}
\usage{
near_prime(n)
}
\description{
this function finds the nearest prime number in a given number. If the number of prime exists at the same distance, print both values.
}
\examples{

near_prime(10)
}
